// server/src/types/comic.ts
export interface Comic {
    month: string
    num: number
    link: string
    img: string
    year: string
    news: string
    save_title: string
    title: string
    alt: string
    transcript: string
    day: string
}

export interface PDLComic {
    creator: string
    content: string
    title: string
    pubDate: string
    link: string
    guid: string
    url: string
}